コンポーネント分割とimport文の変更
