+++ src/components/feature/MemberCard/ProfileImg.tsx
@@ -1,0 +1,14 @@
+import { FC } from "react";
+
+type ProfileImgProps = {
+  displayName: string | undefined;
+  photoURL: string | undefined;
+};
+export const ProfileImg: FC<ProfileImgProps> = ({ displayName, photoURL }) => {
+  return (
+    <div>
+      <img src={photoURL} alt={`${displayName}の画像`} className={`w-12 rounded-full`} />
+      <p className="pt-1 text-xs ">{displayName}</p>
+    </div>
+  );
+};
